+++ tests/tests/core/primitiveset/mod.rs
@@ -14,194 +14,391 @@
 //
 ////////////////////////////////////////////////////////////////////////////////
 
-use tink_core::{primitiveset::Entry, Primitive};
+use tink_core::{primitiveset::Entry, KeyId, MacBoxClone, Primitive};
 use tink_proto::{keyset::Key, KeyStatusType, OutputPrefixType};
 use tink_tests::{new_dummy_key, DummyMac};
 
-fn create_keyset() -> Vec<Key> {
-    let key_id0 = 1234543;
-    let key_id1 = 7213743;
-    let key_id2 = key_id1;
-    let key_id3 = 947327;
-    let key_id4 = 529472;
-    let key_id5 = key_id0;
-    vec![
-        new_dummy_key(key_id0, KeyStatusType::Enabled, OutputPrefixType::Tink),
-        new_dummy_key(key_id1, KeyStatusType::Enabled, OutputPrefixType::Legacy),
-        new_dummy_key(key_id2, KeyStatusType::Enabled, OutputPrefixType::Tink),
-        new_dummy_key(key_id3, KeyStatusType::Enabled, OutputPrefixType::Raw),
-        new_dummy_key(key_id4, KeyStatusType::Enabled, OutputPrefixType::Raw),
-        new_dummy_key(key_id5, KeyStatusType::Enabled, OutputPrefixType::Tink),
-    ]
+fn make_test_key(
+    key_id: KeyId,
+    status: KeyStatusType,
+    output_prefix_type: OutputPrefixType,
+    type_url: &str,
+) -> Key {
+    let mut k = new_dummy_key(key_id, status, output_prefix_type);
+    if let Some(key_data) = &mut k.key_data {
+        key_data.type_url = type_url.to_string();
+    }
+    k
 }
 
 #[test]
-fn test_primitive_set_basic() {
-    let mut ps = tink_core::primitiveset::PrimitiveSet::new();
+fn test_primitiveset_new() {
+    let ps = tink_core::primitiveset::PrimitiveSet::new();
     assert!(ps.primary.is_none());
     assert!(ps.entries.is_empty());
-    // generate test keys
-    let keys = create_keyset();
-    // add all test primitives
-    let mut macs = Vec::with_capacity(keys.len());
-    let mut entries = Vec::with_capacity(keys.len());
-    for i in 0..keys.len() {
-        let mac = Box::new(DummyMac {
-            name: format!("Mac#{}", i),
-        });
-        macs.push(mac);
-        entries.push(ps.add(Primitive::Mac(macs[i].clone()), &keys[i]).unwrap());
+}
+
+fn check_equal(left: &[Entry], right: &[Entry]) {
+    assert_eq!(left.len(), right.len());
+    for i in 0..left.len() {
+        assert_eq!(left[i].key_id, right[i].key_id);
+        assert_eq!(left[i].status, right[i].status);
+        assert_eq!(left[i].prefix_type, right[i].prefix_type);
+        assert_eq!(left[i].prefix, right[i].prefix);
+        assert_eq!(left[i].type_url, right[i].type_url);
+        // `Entry.primitive` is not `PartialEq`, so compare outputs
+        let left_data = if let Primitive::Mac(mac) = &left[i].primitive {
+            mac.compute_mac(&[1, 2, 3, 4, 5]).unwrap()
+        } else {
+            panic!("not a MAC!")
+        };
+        let right_data = if let Primitive::Mac(mac) = &right[i].primitive {
+            mac.compute_mac(&[1, 2, 3, 4, 5]).unwrap()
+        } else {
+            panic!("not a MAC!")
+        };
+        assert_eq!(left_data, right_data);
     }
-
-    // set primary entry
-    let primary_id = 2;
-    ps.primary = Some(entries[primary_id].clone());
-
-    // check raw primitives
-    let raw_ids = vec![keys[3].key_id, keys[4].key_id];
-    let raw_macs = vec![macs[3].clone(), macs[4].clone()];
-    let raw_statuses = vec![
-        KeyStatusType::from_i32(keys[3].status).unwrap(),
-        KeyStatusType::from_i32(keys[4].status).unwrap(),
-    ];
-    let raw_prefix_types = vec![
-        OutputPrefixType::from_i32(keys[3].output_prefix_type).unwrap(),
-        OutputPrefixType::from_i32(keys[4].output_prefix_type).unwrap(),
-    ];
-    let raw_entries = ps.raw_entries();
-    assert!(
-        validate_entry_list(
-            &raw_entries,
-            &raw_ids,
-            &raw_macs,
-            &raw_statuses,
-            &raw_prefix_types
-        ),
-        "raw primitives do not match input"
-    );
-
-    // check tink primitives, same id
-    let tink_ids = vec![keys[0].key_id, keys[5].key_id];
-    let tink_macs = vec![macs[0].clone(), macs[5].clone()];
-    let tink_statuses = vec![
-        KeyStatusType::from_i32(keys[0].status).unwrap(),
-        KeyStatusType::from_i32(keys[5].status).unwrap(),
-    ];
-    let tink_prefix_types = vec![
-        OutputPrefixType::from_i32(keys[0].output_prefix_type).unwrap(),
-        OutputPrefixType::from_i32(keys[5].output_prefix_type).unwrap(),
-    ];
-    let prefix = tink_core::cryptofmt::output_prefix(&keys[0]).unwrap();
-    let tink_entries = ps.entries_for_prefix(&prefix);
-    assert!(
-        validate_entry_list(
-            &tink_entries,
-            &tink_ids,
-            &tink_macs,
-            &tink_statuses,
-            &tink_prefix_types
-        ),
-        "tink primitives do not match the input key"
-    );
-
-    // check another tink primitive
-    let tink_ids = vec![keys[2].key_id];
-    let tink_macs = vec![macs[2].clone()];
-    let tink_statuses = vec![KeyStatusType::from_i32(keys[2].status).unwrap()];
-    let tink_prefix_types = vec![OutputPrefixType::from_i32(keys[2].output_prefix_type).unwrap()];
-    let prefix = tink_core::cryptofmt::output_prefix(&keys[2]).unwrap();
-    let tink_entries = ps.entries_for_prefix(&prefix);
-    assert!(
-        validate_entry_list(
-            &tink_entries,
-            &tink_ids,
-            &tink_macs,
-            &tink_statuses,
-            &tink_prefix_types
-        ),
-        "tink primitives do not match the input key"
-    );
-
-    // check legacy primitives
-    let legacy_ids = vec![keys[1].key_id];
-    let legacy_macs = vec![macs[1].clone()];
-    let legacy_statuses = vec![KeyStatusType::from_i32(keys[1].status).unwrap()];
-    let legacy_prefix_types = vec![OutputPrefixType::from_i32(keys[1].output_prefix_type).unwrap()];
-    let legacy_prefix = tink_core::cryptofmt::output_prefix(&keys[1]).unwrap();
-    let legacy_entries = ps.entries_for_prefix(&legacy_prefix);
-    assert!(
-        validate_entry_list(
-            &legacy_entries,
-            &legacy_ids,
-            &legacy_macs,
-            &legacy_statuses,
-            &legacy_prefix_types
-        ),
-        "legacy primitives do not match the input key"
-    );
 }
 
 #[test]
-fn test_add_with_invalid_input() {
+fn test_primitiveset_add_entries() {
+    let keys = vec![
+        make_test_key(
+            1234543,
+            KeyStatusType::Enabled,
+            OutputPrefixType::Tink,
+            "type.url.1",
+        ),
+        make_test_key(
+            7213743,
+            KeyStatusType::Enabled,
+            OutputPrefixType::Legacy,
+            "type.url.2",
+        ),
+        make_test_key(
+            5294722,
+            KeyStatusType::Enabled,
+            OutputPrefixType::Raw,
+            "type.url.3",
+        ),
+    ];
+    let macs: Vec<DummyMac> = (0..keys.len())
+        .map(|i| DummyMac {
+            name: format!("{}", i),
+        })
+        .collect();
     let mut ps = tink_core::primitiveset::PrimitiveSet::new();
-    let dummy_mac = Box::new(DummyMac {
-        name: "".to_string(),
-    });
-    // unknown prefix type
-    let invalid_key = new_dummy_key(0, KeyStatusType::Enabled, OutputPrefixType::UnknownPrefix);
-    assert!(
-        ps.add(Primitive::Mac(dummy_mac.clone()), &invalid_key)
-            .is_err(),
-        "expect an error when key is invalid"
-    );
 
-    // disabled key
-    let disabled_key = new_dummy_key(0, KeyStatusType::Disabled, OutputPrefixType::UnknownPrefix);
-    assert!(
-        ps.add(Primitive::Mac(dummy_mac), &disabled_key).is_err(),
-        "expect an error when key is disabled"
-    );
+    let got: Vec<Entry> = (0..macs.len())
+        .map(|i| {
+            ps.add(Primitive::Mac(Box::new(macs[i].clone())), &keys[i])
+                .unwrap()
+        })
+        .collect();
+
+    let want = vec![
+        Entry {
+            key_id: 1234543,
+            status: KeyStatusType::Enabled,
+            primitive: Primitive::Mac(Box::new(DummyMac {
+                name: "0".to_string(),
+            })),
+            prefix_type: OutputPrefixType::Tink,
+            type_url: "type.url.1".to_string(),
+            prefix: vec![1, 0, 18, 214, 111],
+        },
+        Entry {
+            key_id: 7213743,
+            status: KeyStatusType::Enabled,
+            primitive: Primitive::Mac(Box::new(DummyMac {
+                name: "1".to_string(),
+            })),
+            prefix_type: OutputPrefixType::Legacy,
+            type_url: "type.url.2".to_string(),
+            prefix: vec![0, 0, 110, 18, 175],
+        },
+        Entry {
+            key_id: 5294722,
+            status: KeyStatusType::Enabled,
+            primitive: Primitive::Mac(Box::new(DummyMac {
+                name: "2".to_string(),
+            })),
+            prefix_type: OutputPrefixType::Raw,
+            type_url: "type.url.3".to_string(),
+            prefix: vec![],
+        },
+    ];
+    check_equal(&got, &want);
 }
 
-fn validate_entry_list(
-    entries: &[Entry],
-    key_ids: &[tink_core::KeyId],
-    macs: &[Box<DummyMac>],
-    statuses: &[KeyStatusType],
-    prefix_types: &[OutputPrefixType],
-) -> bool {
-    if entries.len() != macs.len() {
-        return false;
+#[test]
+fn test_primitiveset_raw_entries() {
+    let keys = vec![
+        make_test_key(
+            1234543,
+            KeyStatusType::Enabled,
+            OutputPrefixType::Tink,
+            "type.url.1",
+        ),
+        make_test_key(
+            7213743,
+            KeyStatusType::Enabled,
+            OutputPrefixType::Legacy,
+            "type.url.2",
+        ),
+        make_test_key(
+            9473277,
+            KeyStatusType::Enabled,
+            OutputPrefixType::Raw,
+            "type.url.3",
+        ),
+        make_test_key(
+            5294722,
+            KeyStatusType::Enabled,
+            OutputPrefixType::Raw,
+            "type.url.4",
+        ),
+    ];
+
+    let macs: Vec<DummyMac> = (0..keys.len())
+        .map(|i| DummyMac {
+            name: format!("Mac#{}", i),
+        })
+        .collect();
+    let mut ps = tink_core::primitiveset::PrimitiveSet::new();
+    for i in 0..macs.len() {
+        ps.add(Primitive::Mac(Box::new(macs[i].clone())), &keys[i])
+            .unwrap();
     }
-    for (i, entry) in entries.iter().enumerate() {
-        if !validate_entry(entry, key_ids[i], &macs[i], &statuses[i], &prefix_types[i]) {
-            return false;
-        }
-    }
-    true
+    let got = ps.raw_entries();
+    let want = vec![
+        Entry {
+            key_id: keys[2].key_id,
+            status: KeyStatusType::from_i32(keys[2].status).unwrap(),
+            prefix_type: OutputPrefixType::from_i32(keys[2].output_prefix_type).unwrap(),
+            type_url: keys[2].key_data.as_ref().unwrap().type_url.clone(),
+            primitive: Primitive::Mac(macs[2].box_clone()),
+            prefix: vec![],
+        },
+        Entry {
+            key_id: keys[3].key_id,
+            status: KeyStatusType::from_i32(keys[3].status).unwrap(),
+            prefix_type: OutputPrefixType::from_i32(keys[3].output_prefix_type).unwrap(),
+            type_url: keys[3].key_data.as_ref().unwrap().type_url.clone(),
+            primitive: Primitive::Mac(macs[3].box_clone()),
+            prefix: vec![],
+        },
+    ];
+    check_equal(&got, &want);
 }
 
-// Compare an entry with the [`DummyMAC`] that was used to create the entry.
-fn validate_entry(
-    entry: &Entry,
-    key_id: tink_core::KeyId,
-    test_mac: &DummyMac,
-    status: &KeyStatusType,
-    output_prefix_type: &OutputPrefixType,
-) -> bool {
-    if entry.key_id != key_id || entry.status != *status || entry.prefix_type != *output_prefix_type
-    {
-        return false;
+#[test]
+fn test_primitiveset_prefixed_entries() {
+    struct TestCase {
+        _tag: &'static str,
+        prefix: Vec<u8>,
+        keys: Vec<Key>,
+        primitives: Vec<DummyMac>,
+        want: Vec<Entry>,
     }
-    if let Primitive::Mac(dummy_mac) = &entry.primitive {
-        let mut data = vec![1, 2, 3, 4, 5];
-        let digest = dummy_mac.compute_mac(&data).unwrap();
-        data.extend_from_slice(test_mac.name.as_bytes());
-        if digest != data {
-            return false;
+    for tc in vec![
+        TestCase {
+            _tag: "legacy Prefix",
+            prefix: vec![0, 0, 18, 214, 111], // LEGACY_PREFIX + 1234543,
+            keys: vec![
+                make_test_key(
+                    1234543,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Legacy,
+                    "type.url.1",
+                ),
+                make_test_key(
+                    7213743,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Tink,
+                    "type.url.2",
+                ),
+            ],
+            primitives: vec![
+                DummyMac {
+                    name: "1".to_string(),
+                },
+                DummyMac {
+                    name: "2".to_string(),
+                },
+            ],
+            want: vec![Entry {
+                key_id: 1234543,
+                status: KeyStatusType::Enabled,
+                primitive: Primitive::Mac(Box::new(DummyMac {
+                    name: "1".to_string(),
+                })),
+                prefix_type: OutputPrefixType::Legacy,
+                type_url: "type.url.1".to_string(),
+                prefix: vec![0, 0, 18, 214, 111],
+            }],
+        },
+        TestCase {
+            _tag: "raw prefix",
+            prefix: vec![],
+            keys: vec![
+                make_test_key(
+                    1234543,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Raw,
+                    "type.url.1",
+                ),
+                make_test_key(
+                    7213743,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Tink,
+                    "type.url.2",
+                ),
+            ],
+            primitives: vec![
+                DummyMac {
+                    name: "1".to_string(),
+                },
+                DummyMac {
+                    name: "2".to_string(),
+                },
+            ],
+            want: vec![Entry {
+                key_id: 1234543,
+                status: KeyStatusType::Enabled,
+                primitive: Primitive::Mac(Box::new(DummyMac {
+                    name: "1".to_string(),
+                })),
+                prefix_type: OutputPrefixType::Raw,
+                type_url: "type.url.1".to_string(),
+                prefix: vec![],
+            }],
+        },
+        TestCase {
+            _tag: "tink prefix  multiple entries",
+            prefix: vec![1, 0, 18, 214, 111], // TINK_PREFIX + 1234543
+            keys: vec![
+                make_test_key(
+                    1234543,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Tink,
+                    "type.url.1",
+                ),
+                make_test_key(
+                    1234543,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Tink,
+                    "type.url.2",
+                ),
+                make_test_key(
+                    1234543,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Raw,
+                    "type.url.3",
+                ),
+                make_test_key(
+                    7213743,
+                    KeyStatusType::Enabled,
+                    OutputPrefixType::Tink,
+                    "type.url.4",
+                ),
+            ],
+            primitives: vec![
+                DummyMac {
+                    name: "1".to_string(),
+                },
+                DummyMac {
+                    name: "2".to_string(),
+                },
+                DummyMac {
+                    name: "3".to_string(),
+                },
+                DummyMac {
+                    name: "4".to_string(),
+                },
+            ],
+            want: vec![
+                Entry {
+                    key_id: 1234543,
+                    status: KeyStatusType::Enabled,
+                    primitive: Primitive::Mac(Box::new(DummyMac {
+                        name: "1".to_string(),
+                    })),
+                    prefix_type: OutputPrefixType::Tink,
+                    type_url: "type.url.1".to_string(),
+                    prefix: vec![1, 0, 18, 214, 111],
+                },
+                Entry {
+                    key_id: 1234543,
+                    status: KeyStatusType::Enabled,
+                    primitive: Primitive::Mac(Box::new(DummyMac {
+                        name: "2".to_string(),
+                    })),
+                    prefix_type: OutputPrefixType::Tink,
+                    type_url: "type.url.2".to_string(),
+                    prefix: vec![1, 0, 18, 214, 111],
+                },
+            ],
+        },
+    ] {
+        let mut ps = tink_core::primitiveset::PrimitiveSet::new();
+        for i in 0..tc.keys.len() {
+            ps.add(Primitive::Mac(tc.primitives[i].box_clone()), &tc.keys[i])
+                .unwrap();
         }
-    } else {
-        panic!("failed to retrieve MAC primitive");
+        let got = ps.entries_for_prefix(&tc.prefix);
+        check_equal(&got, &tc.want);
     }
-    true
+}
+
+#[test]
+fn test_primitiveset_add_with_invalid_input() {
+    let mut ps = tink_core::primitiveset::PrimitiveSet::new();
+    struct TestCase {
+        _tag: &'static str,
+        primitive: DummyMac,
+        key: Key,
+    }
+    for tc in vec![
+        TestCase {
+            _tag: "unknown prefix type",
+            primitive: DummyMac {
+                name: "".to_string(),
+            },
+            key: make_test_key(
+                0,
+                KeyStatusType::Enabled,
+                OutputPrefixType::UnknownPrefix,
+                "type.url.1",
+            ),
+        },
+        TestCase {
+            _tag: "disabled key",
+            primitive: DummyMac {
+                name: "".to_string(),
+            },
+            key: make_test_key(
+                0,
+                KeyStatusType::Disabled,
+                OutputPrefixType::Tink,
+                "type.url.1",
+            ),
+        },
+        TestCase {
+            _tag: "nil key_data",
+            primitive: DummyMac {
+                name: "".to_string(),
+            },
+            key: Key {
+                key_data: None,
+                status: KeyStatusType::Enabled as i32,
+                output_prefix_type: OutputPrefixType::Tink as i32,
+                key_id: 0,
+            },
+        },
+    ] {
+        let result = ps.add(Primitive::Mac(Box::new(tc.primitive)), &tc.key);
+        assert!(result.is_err());
+    }
 }